fix ERC20IdlePCVDeposit   - [x] Name convention IdlePCVDeposit -> ERC20IdlePCVDeposit  - [x] Wraping safeTransfer call in `if(to !=address(this))`  - [x] Using `Constants` library  - [x] Creating separate a file `IIdleToken` interface  - [x] Testing `resistantBalanceAndFei()` function
