+++ contracts/pcv/idle/ERC20IdlePCVDeposit.sol
@@ -3,39 +3,18 @@ pragma solidity ^0.8.4;
 
 import "../PCVDeposit.sol";
 import "../../refs/CoreRef.sol";
+import "../../Constants.sol";
+import "./IIdleToken.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
-interface IIdleToken {
-    /* ==== ERC20 Methods ==== */
-    function balanceOf(address user) external view returns (uint256 amount);
-
-    /* ==== Methods unique to IdleToken ==== */
-    function token() external view returns (address underlying);
-
-    function mintIdleToken(
-        uint256 _amount,
-        bool _skipWholeRebalance,
-        address _referral
-    ) external returns (uint256 mintedTokens);
-
-    function redeemIdleToken(uint256 _amount) external returns (uint256 redeemedTokens);
-
-    function tokenPriceWithFee(address user) external view returns (uint256 priceWFee);
-}
-
 /// @title ERC-20 implementation for a Idle PCV Deposit
-/// @author Fei Protocol
-contract IdlePCVDeposit is PCVDeposit {
-    uint256 private constant EXCHANGE_RATE_SCALE = 1e18;
-
+/// @author massun-onibakuchi
+contract ERC20IdlePCVDeposit is PCVDeposit {
     /// @notice the token underlying the IdleToken
     IERC20 public token;
 
     /// @notice the IdleToken
     IIdleToken public idleToken;
-
-    /// @notice refferal address for the IdleToken
-    address public referral;
 
     /// @notice Idle ERC20 PCV Deposit constructor
     /// @param _core Fei Core for reference
@@ -44,14 +23,12 @@ contract IdlePCVDeposit is PCVDeposit {
     constructor(
         address _core,
         IIdleToken _idleToken,
-        IERC20 _token,
-        address _referral
+        IERC20 _token
     ) CoreRef(_core) {
         require(_idleToken.token() == address(_token), "IdlePCVDeposit: Not the same token");
 
         token = _token;
         idleToken = _idleToken;
-        referral = _referral; // Can be zero address
     }
 
     /// @notice deposit ERC-20 tokens to Idle
@@ -60,7 +37,7 @@ contract IdlePCVDeposit is PCVDeposit {
 
         token.approve(address(idleToken), amount);
 
-        idleToken.mintIdleToken(amount, false, referral);
+        idleToken.mintIdleToken(amount, false, address(0));
 
         emit Deposit(msg.sender, amount);
     }
@@ -75,23 +52,33 @@ contract IdlePCVDeposit is PCVDeposit {
         // Because of rounding, the actual amount of underlying received may differ.
         // So query the actual amount received.
         uint256 balanceBefore = token.balanceOf(address(this));
-        idleToken.redeemIdleToken((amountUnderlying * EXCHANGE_RATE_SCALE) / priceWithFee);
+        idleToken.redeemIdleToken((amountUnderlying * Constants.ETH_GRANULARITY) / priceWithFee);
 
         amountUnderlying = token.balanceOf(address(this)) - balanceBefore;
 
-        SafeERC20.safeTransfer(token, to, amountUnderlying);
+        // Fei may withdraw to self
+        if (to != address(this)) {
+            SafeERC20.safeTransfer(token, to, amountUnderlying);
+        }
         emit Withdrawal(msg.sender, to, amountUnderlying);
     }
 
-    /// @notice returns total balance of PCV in the Deposit excluding the FEI
-    /// @dev returns stale values from Idle if the market hasn't been updated
+    // @todo check for manipulation resistance (flashloans, etc)
+    /// @notice return total balance of "balanceReportedIn" PCV token in the Deposit
+    /// @dev return the balance by using current idleToken price based on net asset value and totalSupply
     function balance() public view override returns (uint256) {
         uint256 priceWithFee = idleToken.tokenPriceWithFee(address(this));
-        return (idleToken.balanceOf(address(this)) * priceWithFee) / EXCHANGE_RATE_SCALE;
+        return (idleToken.balanceOf(address(this)) * priceWithFee) / Constants.ETH_GRANULARITY;
     }
 
     /// @notice display the related token of the balance reported
     function balanceReportedIn() public view override returns (address) {
         return address(token);
+    }
+
+    function resistantBalanceAndFei() public view override returns (uint256, uint256) {
+        uint256 _balance = balance();
+        if (address(token) == address(fei())) return (0, _balance);
+        else return (_balance, 0);
     }
 }
