+++ contracts/pcv/idle/IIdleToken.sol
@@ -1,0 +1,49 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+pragma solidity ^0.8.4;
+
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+
+/// @title Interface for the Idle best-yield strategy
+interface IIdleToken is IERC20 {
+    /**
+     * @return underlying token address
+     */
+    function token() external view returns (address underlying);
+
+    /**
+     * Used to mint IdleTokens, given an underlying amount (eg. DAI).
+     * This method triggers a rebalance of the pools if needed
+     * NOTE: User should 'approve' _amount of tokens before calling mintIdleToken
+     * NOTE 2: this method can be paused
+     *
+     * @param _amount : amount of underlying token to be lended
+     * @param _skipWholeRebalance : flag for skipping rebalance for lower gas price
+     * @param _referral : referral address
+     * @return mintedTokens : amount of IdleTokens minted
+     */
+    function mintIdleToken(
+        uint256 _amount,
+        bool _skipWholeRebalance,
+        address _referral
+    ) external returns (uint256 mintedTokens);
+
+    /**
+     * Here we calc the pool share one can withdraw given the amount of IdleToken they want to burn
+     * This method triggers a rebalance of the pools if needed
+     * NOTE: If the contract is paused or iToken price has decreased one can still redeem but no rebalance happens.
+     * NOTE 2: If iToken price has decresed one should not redeem (but can do it) otherwise he would capitalize the loss.
+     *         Ideally one should wait until the black swan event is terminated
+     *
+     * @param _amount : amount of IdleTokens to be burned
+     * @return redeemedTokens : amount of underlying tokens redeemed
+     */
+    function redeemIdleToken(uint256 _amount) external returns (uint256 redeemedTokens);
+
+    /**
+     * @notice IdleToken price for a specific user considering fees, in underlying.
+     * this is useful when you need to redeem exactly X underlying
+     * @param user address
+     * @return priceWFee : IdleToken price with fees
+     */
+    function tokenPriceWithFee(address user) external view returns (uint256 priceWFee);
+}
